Add Webpack back in
Unfortunately, it seems like the changes that were supposed to
be coming with elm-reactor in 0.18 are not as far reaching as
it seemed they would be. There's a bug in the reactor where if
you try and install your elm packages all at once, it has a
runtime error and crashes. So, aside from some prototyping where
you install the core packages, then other stuff, the reactor is
honestly kind of useless.

I really want that exportable history. But it's too hard to get it
working in a way that anyone else would be able to easily pick it
up. And I'm not a huge fan of how webpack is so poorly documented
but it's the best option right now.