Close #3: Add rudamentary chassis Designer
This is **really** basic, but it's all that exists so far as a mek
is concerned for the moment. The main Elm file is pretty big, and
I'll probably break it up later.

Also, there's no styling on this. I haven't really given thought to
what it looks like just yet, only that it is functional for the moment.
I think coming up with a good UI will come after I know that everything
is hooked up properly.