Add installable Engines
It's worth noting here that I'm realizing that something is not right with how
this is coming together. It has to do with the issue that two very prominent
components; the Engine and the Gyroscope, are both concepts, and individual
components in the Mek. And I'm not entierly sure how to model that.

They aren't strictly a pure result of interrogating some other information in
the Mek, because the components themselves may be damaged. Which means that the
components are somehow stateful in the Mek. I'm not sure exactly how to go about
reflecting this just yet.